Add socket tracer to send all events to a socket
This will be useful to use many machines for profiling heavy
applications. In addition to the machine being traced, we can
have one machine per number of tracers in the pool size. We
can also make use of different network interfaces to duplicate
the bandwidth.

We might want to allow compression in the future. I am not
100% convinced it is necessary yet. Experiments will tell.

This commit does not include the documentation for the feature.