Improve error handling in file reader; increase event size field
The increased event size field is to fix issues reading very large
events, we may have some when messages are involved for example.

A separate commit will add the truncating of large events so this
increase is really "just in case" as when it is too small the
reading of the event fails completely.