Fix return_to event handling
The functions returning were not properly handled
in some cases. The current function always returns, but
some tail calls may also end at the same time. Finally
we return to a function on the stack (or nothing, depending
on the trace rules). All three must be updated differently.
In particular, the calculations for the function on the
stack must resume counting for self from the time we return
to this function.

When storing the result in our state we were not properly
aggregating wait times. This led to functions that do not
do anything (for example a simple receive statement) to
have large active time. Now the wait time is properly
counted, at least in the case tested.

There might still be some issues with incl/self around
the wait times or the number of times we wait. This should
be investigated as part of a different commit.