Improve wait_count calculations
It is now confirmed correct in at least some cases (where we
have a common function as a starting point). It may not be
correct when traces have more than one function as a starting
point (for example when tracing on a specific module with
function calls being done in/out of this module). A new test
will be needed for this.

Note that even despite `qcachegrind` showing the correct
count there are still invalid values in the produced
callgrind.out files (but these values are seemingly not
used in the particular case that I fixed).